+++ Source/Aggregates/AggregateRoot.cs
@@ -4,6 +4,8 @@
 using System;
 using System.Collections.Generic;
 using System.Threading;
+using System.Threading.Tasks;
+using Dolittle.SDK.Aggregates.Internal;
 using Dolittle.SDK.Artifacts;
 using Dolittle.SDK.Events;
 using Dolittle.SDK.Events.Store;
@@ -13,19 +15,28 @@ namespace Dolittle.SDK.Aggregates;
 /// <summary>
 /// Represents the aggregate root.
 /// </summary>
-public class AggregateRoot
+public abstract class AggregateRoot
 {
-    readonly IList<AppliedEvent> _appliedEvents;
-
+    readonly List<AppliedEvent> _appliedEvents = new();
+    EventSourceId? _eventSourceId;
     /// <summary>
     /// Initializes a new instance of the <see cref="AggregateRoot"/> class.
     /// </summary>
     /// <param name="eventSourceId">The <see cref="Events.EventSourceId" />.</param>
-    public AggregateRoot(EventSourceId eventSourceId)
+    [Obsolete("This base constructor is deprecated and only used to set the EventSourceId property so that it could be used in the constructor")]
+    protected AggregateRoot(EventSourceId eventSourceId)
+        : this()
     {
         EventSourceId = eventSourceId;
+    }
+    /// <summary>
+    /// Initializes a new instance of the <see cref="AggregateRoot"/> class.
+    /// </summary>
+    protected AggregateRoot()
+    {
+        AggregateRootId = this.GetAggregateRootId();
         Version = AggregateRootVersion.Initial;
-        _appliedEvents = new List<AppliedEvent>();
+        IsStateless = this.IsStateless();
     }
 
     /// <summary>
@@ -34,14 +45,36 @@ public class AggregateRoot
     public AggregateRootVersion Version { get; private set; }
 
     /// <summary>
+    /// Gets the <see cref="Events.AggregateRootId"/>.
+    /// </summary>
+    public AggregateRootId AggregateRootId { get; }
+
+    /// <summary>
     /// Gets the <see cref="Events.EventSourceId" /> that the <see cref="AggregateRoot" /> applies events to.
     /// </summary>
-    public EventSourceId EventSourceId { get; }
+    [Obsolete("This will eventually be marked as internal. If you need to know the event source id in the aggregate root then include it in the constructor and keep it as a field")]
+    public EventSourceId EventSourceId
+    {
+        get => _eventSourceId ?? throw new EventSourceIdOnAggregateRootNotReady(GetType());
+        internal set
+        {
+            if (_eventSourceId is null)
+            {
+                _eventSourceId = value;
+            }
+            else if (_eventSourceId.Value != value.Value)
+            {
+                throw new CannotChangeEventSourceIdForAggregateRoot(GetType(), _eventSourceId, value);
+            }
+        }
+    }
 
     /// <summary>
     /// Gets the <see cref="IEnumerable{T}" /> of applied events to commit.
     /// </summary>
     public IEnumerable<AppliedEvent> AppliedEvents => _appliedEvents;
+    
+    bool IsStateless { get; }
 
     /// <summary>
     /// Apply the event to the <see cref="AggregateRoot" /> so that it will be committed to the <see cref="IEventStore" />
@@ -124,25 +157,47 @@ public class AggregateRoot
         => Apply(@event, eventType, true);
 
     /// <summary>
-    /// Re-apply events from the Event Store.
+    /// Rehydrates the aggregate root with the <see cref="CommittedAggregateEvents"/> for this aggregate.
     /// </summary>
-    /// <param name="events">Sequence that contains the events to re-apply.</param>
-    public virtual void ReApply(CommittedAggregateEvents events)
+    /// <param name="batches">The <see cref="IAsyncEnumerator{T}"/> batches of <see cref="CommittedAggregateEvents"/> to rehydrate with.</param>
+    /// <param name="cancellationToken">The <see cref="CancellationToken"/> used for cancelling the rehydration.</param>
+    public async Task Rehydrate(IAsyncEnumerable<CommittedAggregateEvents> batches, CancellationToken cancellationToken)
     {
-        ThrowIfEventWasAppliedToOtherEventSource(events);
-        ThrowIfEventWasAppliedByOtherAggregateRoot(events);
-
-        foreach (var @event in events)
+        var hasBatches = false;
+        var expectedVersion = AggregateRootVersion.Initial;
+        await foreach (var batch in batches.WithCancellation(cancellationToken))
         {
-            ThrowIfAggreggateRootVersionIsOutOfOrder(@event);
-            Version++;
-            if (!this.IsStateless()) InvokeOnMethod(@event.Content);
+            expectedVersion = batch.AggregateRootVersion;
+            hasBatches = true;
+            ThrowIfEventWasAppliedToOtherEventSource(batch);
+            ThrowIfEventWasAppliedByOtherAggregateRoot(batch);
+            if (IsStateless)
+            {
+                break;
+            }
+            foreach (var @event in batch)
+            {
+                Version = @event.AggregateRootVersion + 1;
+                InvokeOnMethod(@event.Content);
+            }
         }
+        if (!hasBatches)
+        {
+            throw new NoCommittedAggregateEventsBatches(AggregateRootId, EventSourceId);
+        }
+        
+        Version = expectedVersion;
+        cancellationToken.ThrowIfCancellationRequested();
     }
 
-    void Apply(object @event, EventType eventType, bool isPublic)
+
+    void Apply(object @event, EventType? eventType, bool isPublic)
     {
-        if (@event == null) throw new EventContentCannotBeNull();
+        if (@event == null)
+        {
+            throw new EventContentCannotBeNull();
+        }
+
         _appliedEvents.Add(new AppliedEvent(@event, eventType, isPublic));
         Version++;
         InvokeOnMethod(@event);
@@ -156,19 +211,20 @@ public class AggregateRoot
         }
     }
 
-    void ThrowIfAggreggateRootVersionIsOutOfOrder(CommittedAggregateEvent @event)
-    {
-        if (@event.AggregateRootVersion != Version) throw new AggregateRootVersionIsOutOfOrder(@event.AggregateRootVersion, Version);
-    }
-
     void ThrowIfEventWasAppliedByOtherAggregateRoot(CommittedAggregateEvents events)
     {
         var aggregateRootId = this.GetAggregateRootId();
-        if (events.AggregateRoot != this.GetAggregateRootId()) throw new EventWasAppliedByOtherAggregateRoot(events.AggregateRoot, aggregateRootId);
+        if (events.AggregateRoot != this.GetAggregateRootId())
+        {
+            throw new EventWasAppliedByOtherAggregateRoot(events.AggregateRoot, aggregateRootId);
+        }
     }
 
     void ThrowIfEventWasAppliedToOtherEventSource(CommittedAggregateEvents events)
     {
-        if (events.EventSource != EventSourceId) throw new EventWasAppliedToOtherEventSource(events.EventSource, EventSourceId);
+        if (events.EventSource != EventSourceId)
+        {
+            throw new EventWasAppliedToOtherEventSource(events.EventSource, EventSourceId);
+        }
     }
 }