+++ Source/Aggregates/Internal/IAggregateRoots.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
 using Dolittle.SDK.Async;
 using Dolittle.SDK.Events;
 
@@ -15,8 +16,9 @@ public interface IAggregateRoots
     /// Tries to get an <typeparamref name="TAggregate"/> instance for the specified <see cref="EventSourceId"/>.
     /// </summary>
     /// <param name="eventSourceId">The <see cref="EventSourceId"/> of the aggregate root instance to create.</param>
+    /// <param name="serviceProvider">The tenant scoped <see cref="IServiceProvider"/>.</param>
     /// <typeparam name="TAggregate"><see cref="AggregateRoot"/> type.</typeparam>
     /// <returns>A <see cref="Try{TResult}"/> with the <typeparamref name="TAggregate"/>.</returns>
-    Try<TAggregate> TryGet<TAggregate>(EventSourceId eventSourceId)
+    Try<TAggregate> TryGet<TAggregate>(EventSourceId eventSourceId, IServiceProvider serviceProvider)
         where TAggregate : AggregateRoot;
 }