+++ Source/DependencyInversion/ServiceScopeFactory.cs
@@ -34,6 +34,9 @@ public class ServiceScopeFactory : IServiceScopeFactory
         return new TenantServiceScope(
             rootScope,
             _lifetimeScope.BeginLifetimeScope(builder =>
-                builder.RegisterSource(new UnknownServiceOnTenantContainerRegistrationSource(rootScope.ServiceProvider, _lifetimeScope.ComponentRegistry.Registrations, false))));
+                builder.RegisterSource(new UnknownServiceOnTenantContainerRegistrationSource(rootScope.ServiceProvider, false)
+                {
+                    Registrations = _lifetimeScope.ComponentRegistry.Registrations
+                })));
     }
 }