+++ Specifications/Events/Store/Converters/for_AggregateEventToSDKConverter/when_converting_committed_aggregate_events/and_content_is_null.cs
@@ -33,6 +33,7 @@ public class and_content_is_null : given.committed_aggregate_events_and_a_conver
             EventSourceId = event_source.Value,
             AggregateRootId = aggregate_root_id.ToProtobuf(),
             AggregateRootVersion = aggregate_root_version,
+            CurrentAggregateRootVersion = aggregate_root_version,
             Events = { committed_aggregate_event }
         };