+++ CHANGELOG.md
@@ -1,3 +1,75 @@
+# [18.1.1] - 2022-10-31 [PR: #164](https://github.com/dolittle/DotNET.SDK/pull/164)
+## Summary
+
+A recently added bug in the aggregate root that would occur for aggregate roots where the aggregate committed events that was not used to manipulate the state of the aggregate root.
+
+### Fixed
+
+- Set the correct aggregate root version at the end of rehydration.
+
+
+# [18.1.0] - 2022-10-26 [PR: #163](https://github.com/dolittle/DotNET.SDK/pull/163)
+## Summary
+
+Enables dependency injection for aggregate roots
+
+### Added
+
+- Tenant scoped dependencies will now be injected in the aggregate roots
+
+### Deprecated
+
+- AggregateRoot base constructor with event source id is no longer necessary
+
+
+# [18.0.1] - 2022-10-20 [PR: #162](https://github.com/dolittle/DotNET.SDK/pull/162)
+## Summary
+
+Fixes for a problem with the tenant scoped containers where it could not resolve tenant scoped dependencies registered as callbacks.
+
+### Fixed
+
+- Tenant Scoped Containers
+
+
+# [18.0.0] - 2022-9-15 [PR: #160](https://github.com/dolittle/DotNET.SDK/pull/160)
+## Summary
+
+Adds a new method to the `EventStore` for fetching committed aggregate events filtered by event types. This allows us to change the rehydration of aggregates to be much more effective by just fetching the committed aggregate events for an aggregate that are relevant to the rehydration (meaning that there are an `On` method for that event type). This can have a significant impact on the performance of aggregates that have many events, but few state changes or are completely stateless. 
+
+### Added
+
+- `FetchForAggregate` that takes in a list event types used for filtering
+- `FetchStreamForAggregate` fetches a stream of committed aggregate event batches
+
+### Changed
+
+- Rehydration logic of aggregate roots. It now only fetches the committed aggregate events that are relevant
+- `AggregateRootVersion` on the `CommittedAggregateEvents` now represents the current aggregate root version of the aggregate root, not the version of the last committed aggregate event
+- Minor version of Contracts meaning that this version of the SDK is only compatible with version `>= 8.5.0` of the Runtime
+
+
+### Fixed
+
+- Some users could experience exceptions when performing actions on aggregate roots that had lots of events or big events due to the protobuf messages being too big. This should be fixed now since the internals of fetching aggregate roots are now implemented using streaming and batching.
+
+
+# [17.2.3] - 2022-9-9 [PR: #161](https://github.com/dolittle/DotNET.SDK/pull/161)
+### Fixed
+
+* Fixes bug in `WithOpenTelemetrySettings`
+
+
+# [17.2.2] - 2022-8-26 [PR: #158](https://github.com/dolittle/DotNET.SDK/pull/158)
+## Summary
+
+We have experienced applications that were crashing due to grpc messages growing too big due to the volume of events for an aggregate root. As a quick workaround for this we can increase the maximum size of the grpc messages. In the long term a better solution will be to implement streaming and batch mechanisms.
+
+### Changed
+
+- The max size of grpc messages to 32 mb
+
+
 # [17.2.1] - 2022-8-10 [PR: #156](https://github.com/dolittle/DotNET.SDK/pull/156)
 ## Summary